Rewrite updateArea() method using ImageData
This new version creates an ImageData which it fills by
quckly translating the data in displayColours and
displayPixels to RGBA 4-byte grid of pixels in ImageData.

This is much faster than doing direct drawing to the canvas.

Also this ensures we can't violate the way ZX Spectrum
renders colours in blocks (the infamous colour clash).

* Remove NORMAL_COLOURS and BRIGHT_COLOURS. They are redundant now.